adapt canBeReturnedForDays to int with different semantic

+++ src/main/kotlin/org/misarch/returns/event/model/ProductVariantVersionDTO.kt
@@ -10,5 +10,5 @@ import java.util.*
  */
 data class ProductVariantVersionDTO(
     val id: UUID,
-    val canBeReturnedForDays: Double?,
+    val canBeReturnedForDays: Int?,
 )
